[优化]将截图检测移动出来，方便别的代码复用
