+++ common/screenshot.py
@@ -1,0 +1,54 @@
+# -*- coding: utf-8 -*-
+"""
+手机屏幕截图的代码
+"""
+import subprocess
+import os
+import sys
+from PIL import Image
+
+
+# SCREENSHOT_WAY是截图方法，经过check_screenshot后，会自动递减，因此不需修改
+SCREENSHOT_WAY = 3
+
+
+def pull_screenshot():
+    """
+    获取屏幕接入，目前有 0 1 2 3 四种方法，未来添加新的平台监测
+    方法时，可根据效率及适用性由高到低排序
+    """
+    global SCREENSHOT_WAY
+    if 1 <= SCREENSHOT_WAY <= 3:
+        process = subprocess.Popen(
+            'adb shell screencap -p',
+            shell=True, stdout=subprocess.PIPE)
+        binary_screenshot = process.stdout.read()
+        if SCREENSHOT_WAY == 2:
+            binary_screenshot = binary_screenshot.replace(b'\r\n', b'\n')
+        elif SCREENSHOT_WAY == 1:
+            binary_screenshot = binary_screenshot.replace(b'\r\r\n', b'\n')
+        f = open('autojump.png', 'wb')
+        f.write(binary_screenshot)
+        f.close()
+    elif SCREENSHOT_WAY == 0:
+        os.system('adb shell screencap -p /sdcard/autojump.png')
+        os.system('adb pull /sdcard/autojump.png .')
+
+
+def check_screenshot():
+    """
+    检查获取截图的方式
+    """
+    global SCREENSHOT_WAY
+    if os.path.isfile('autojump.png'):
+        os.remove('autojump.png')
+    if SCREENSHOT_WAY < 0:
+        print('暂不支持当前设备')
+        sys.exit()
+    pull_screenshot()
+    try:
+        Image.open('./autojump.png').load()
+        print('采用方式 {} 获取截图'.format(SCREENSHOT_WAY))
+    except Exception:
+        SCREENSHOT_WAY -= 1
+        check_screenshot()
